dep-updates: Update how we use pdfjs in cypress
